refactor(@angular/cli): remove duplicate npmrc logic
