+++ packages/angular/cli/utilities/package-json.ts
